requirements.txt
Flask
requests
debugpy 
google-cloud-bigquery

